% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gauss_MCMC.R
\name{Gauss_model}
\alias{Gauss_model}
\title{Function for the Gaussian fused model}
\usage{
Gauss_model(
  y1,
  y2,
  X = NULL,
  G,
  nIter = 5000,
  beta_thres = 10,
  nug_sig1 = 0.05,
  nug_sig2 = 0.1,
  scale_by_sigma = FALSE,
  verbose = "TRUE"
)
}
\arguments{
\item{y1}{is the expression vector of gene 1, gene 1 will be regressed on gene 2}

\item{y2}{is the expression vector of gene 2}

\item{X}{is the matrix of additional cell-level covariates}

\item{nIter}{is the number of MCMC iterations}

\item{beta_thres}{is a hard threshold on the absolute values of beta's}

\item{nug_sig1}{is the additional precision on beta0's}

\item{nug_sig2}{is the additional precision on beta1's}

\item{scale_by_sigma}{is TRUE or FALSE, whether to scale the beta's by variance of y or not}

\item{verbose}{is TRUE or FALSE, to show or suppress progression updates}
}
\value{
It returns the estimated beta's, sigma, shrinkage parameters, and deviance
}
\description{
Function for the Gaussian fused model
}
