% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Additional_functions.R
\name{kernel_mat}
\alias{kernel_mat}
\title{Function for kernel matrix construction based on the distance matrix}
\usage{
kernel_mat(x, l, type = "exponential")
}
\arguments{
\item{x}{is a dist object or an Euclidean distance matrix}

\item{l}{is the lengthscale, controlling the exponential decay}

\item{type}{denotes the type of kernel function, exponential or radial basis function (RBF)}
}
\value{
To visualize the MST
}
\description{
Function for kernel matrix construction based on the distance matrix
}
