% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result_generator.R
\name{local_res}
\alias{local_res}
\title{Perform local significance test}
\usage{
local_res(Beta, summary = "median")
}
\arguments{
\item{Beta}{is a dataframe of dimension (nIter x N) with MCMC beta samples, for N cells/locations}

\item{summary}{is the type of posterior summary measure to use, either "median" or "mean"}
}
\value{
a data.frame of local beta estimates (N rows for N locations), local p-values, and 
Geweke convergence-diagnostic statistic for each local beta.
}
\description{
Perform local significance test
}
