% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result_generator.R
\name{MCMC_plot}
\alias{MCMC_plot}
\title{MCMC convergence plot}
\usage{
MCMC_plot(x)
}
\arguments{
\item{x}{is a vector of MCMC samples, of dimension nIter x 1}
}
\value{
a plot of MCMC samples with posterior mean in color green
}
\description{
MCMC convergence plot
}
