% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mela_coords}
\alias{mela_coords}
\title{mela_coords dataset}
\format{
A data frame with N = 293 rows (spots) and 3 columns (xy coordinates and celltypes):
\describe{
  \item{x}{x co-ordinate of the cell}
  \item{y}{y co-ordinate of the cell}
  \item{celltypes}{Celltypes detected by the RCTD software based on the gene expression profile}
}
}
\source{
\url{https://aacrjournals.org/cancerres/article/78/20/5970/631815/Spatially-Resolved-Transcriptomics-Enables}
}
\usage{
data(mela_coords)
}
\description{
The location coordinates of the melanoma dataset
}
\keyword{datasets}
