% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Additional_functions.R
\name{MST_construction}
\alias{MST_construction}
\title{Function for MST construction}
\usage{
MST_construction(x, type = "random")
}
\arguments{
\item{x}{is a dist object or an Euclidean distance matrix}

\item{type}{of the MST, random or direct}
}
\value{
It returns the MST inside a list
}
\description{
Function for MST construction
}
