% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NB_MCMC_spam_update.R
\name{NB_model}
\alias{NB_model}
\title{Function for the NB fused model}
\usage{
NB_model(
  y1,
  y2,
  X = NULL,
  G,
  nIter = 5000,
  beta_thres = 10,
  nug_sig1 = 0.05,
  nug_sig2 = 0.1,
  which.r.sampler = "CRT",
  verbose = "TRUE"
)
}
\arguments{
\item{y1}{is the expression vector of gene 1, gene 1 will be regressed on gene 2}

\item{y2}{is the expression vector of gene 2}

\item{X}{is the matrix of additional cell-level covariates}

\item{nIter}{is the number of MCMC iterations}

\item{beta_thres}{is a hard threshold on the absolute values of beta's}

\item{nug_sig1}{is the additional precision on beta0's}

\item{nug_sig2}{is the additional precision on beta1's}

\item{which.r.sampler}{is the choice of sampling r, either CRT or MH}

\item{verbose}{is TRUE or FALSE, to show or suppress progression updates}
}
\value{
It returns the estimated beta's, r, shrinkage parameters, and deviance
}
\description{
Function for the NB fused model
}
