% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result_generator.R
\name{plot_local_comparisons}
\alias{plot_local_comparisons}
\title{Function for comparing HS vs, ICAR local coefficients and p-values}
\usage{
plot_local_comparisons(
  L_results_first = L_results,
  L_results_second = L_results_ICAR,
  point_alpha = 0.6,
  point_size = 1.6
)
}
\arguments{
\item{L_results_first}{is the output of local_res() function on the slope coefficients from the first model}

\item{L_results_second}{is the output of local_res() function on the slope coefficients from the second model}
}
\value{
The kernel covariance matrix of dimension N x N, between N locations
}
\description{
Function for comparing HS vs, ICAR local coefficients and p-values
}
