% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Additional_functions.R
\name{print_mst}
\alias{print_mst}
\title{Function for plotting the MST}
\usage{
print_mst(mst_out, coords, title.hjust = 0.5, title.size = 10, node.size = 1)
}
\arguments{
\item{mst_out}{is the list output from function MST_construction}

\item{coords}{is the 2D matrix of coordinates}

\item{title.hjust}{denotes the position of the plot title, default is centered}

\item{title.size}{denotes the size of the plot title}

\item{node.size}{denotes the size of each node/location of the MST}
}
\value{
To visualize the MST
}
\description{
Function for plotting the MST
}
