% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceBF_main.R
\name{SpaceBF}
\alias{SpaceBF}
\title{Main parent function for fitting different models}
\usage{
SpaceBF(
  y1,
  y2,
  X = NULL,
  G,
  which.model = "NB",
  nIter = 5000,
  beta_thres = 10,
  nug_sig = NULL,
  which.r.sampler = "CRT",
  scale_by_sigma = FALSE,
  verbose = "TRUE"
)
}
\arguments{
\item{y1}{is the expression vector of gene 1, gene 1 will be regressed on gene 2}

\item{y2}{is the expression vector of gene 2}

\item{X}{is the matrix of additional cell-level covariates}

\item{which.model}{denotes the model to be used, "NB" or "Gaussian"}

\item{nIter}{is the number of MCMC iterations}

\item{beta_thres}{is a hard threshold on the absolute values of beta's}

\item{nug_sig}{is the additional precision parameters on beta's, either NULL or a vector of size two}

\item{which.r.sampler}{is the choice of sampling r in the NB model, either CRT or MH}

\item{scale_by_sigma}{is TRUE or FALSE for the Gaussian model, whether to scale the beta's by variance of y or not}

\item{verbose}{is TRUE or FALSE, to show or suppress progression updates}
}
\value{
It returns the estimated beta's, r or sigma, shrinkage parameters, and deviance
}
\description{
Main parent function for fitting different models
}
