% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SpaceBF_cSCC}
\alias{SpaceBF_cSCC}
\title{SpaceBF_cSCC dataset}
\format{
A list of three data frames
}
\source{
\url{https://www.cell.com/cell/fulltext/S0092-8674(20)30672-3}
}
\usage{
data(SpaceBF_cSCC)
}
\description{
The full cSCC dataset: gene expression, xy-coordinates, and keratin pairs
}
\keyword{datasets}
