% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SpaceBF_melanoma}
\alias{SpaceBF_melanoma}
\title{SpaceBF_melanoma dataset}
\format{
A list with components:
\describe{
  \item{gene_data_mela}{gene expression data}
  \item{coords_data_mela}{xy co-ordinates}
  \item{LR_pairs}{161 ligand–receptor pairs}
}
}
\source{
\url{https://aacrjournals.org/cancerres/article/78/20/5970/631815/Spatially-Resolved-Transcriptomics-Enables}
}
\usage{
data(SpaceBF_melanoma)
}
\description{
The full melanoma dataset: gene expression, xy-coordinates, and LR pairs
}
\keyword{datasets}
