% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result_generator.R
\name{global_res}
\alias{global_res}
\title{Perform global significance test}
\usage{
global_res(Beta, summary = "median", local.p.ths = 0.95)
}
\arguments{
\item{Beta}{is a dataframe of dimension (nIter x N) with MCMC beta samples, for N cells/locations}

\item{summary}{is the type of posterior summary measure to use, either "median" or "mean"}

\item{local.p.ths}{is the size of the credible interval for location-specific tests}
}
\value{
a list of global beta estimate (mean or median across locations),
global p-value, number of locations where local CI does not contain 0 (i.e., significant), 
and Geweke convergence-diagnostic statistic for the global beta
}
\description{
Perform global significance test
}
