% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result_generator.R
\name{local_credible}
\alias{local_credible}
\title{Local-level or cell-specific credible interval}
\usage{
local_credible(Beta, p.ths = 0.95)
}
\arguments{
\item{Beta}{is a dataframe of dimension (nIter x N) with MCMC beta samples, for N cells/locations}

\item{p.ths}{is the size of the credible interval}
}
\value{
a heatmap object from the pheatmap package.
}
\description{
Local-level or cell-specific credible interval
}
